+++ src/Precept.Analyzers/DiagnosticCoverageAllowLists.cs
@@ -200,7 +200,9 @@ internal static class DiagnosticCoverageAllowLists
         "UnterminatedInterpolation",
         "UnterminatedStringLiteral",
         "UnterminatedTypedConstant",
+        "UninitializedCrossFieldReadInInitialAssignment",
         "UninitializedFieldReadInInitialAssignment",
+        "MaterializedFieldSelfReference",
         "WritableOnEventArg",
     };
 }